requirements.txt
opencv_python==4.5.1.48
scikit_image==0.17.2
numpy==1.19.5
tensorflow_gpu==2.4.1
matplotlib==3.3.4
pandas==1.1.5
typing_extensions==3.7.4.3
Pillow==8.3.2
skimage==0.0
tensorflow==2.6.0
